requirements.txt
absl-py==0.12.0
astor==0.8.1
boto3==1.17.60
botocore==1.20.60
imutils==0.5.4
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
mock==4.0.3
numpy==1.19.3
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.1.2
protobuf==3.15.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
tensorflow==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
matplotlib
mtcnn
unidecode
os
hausdorff==0.2.6
numba==0.56.4